requirements.txt
python-telegram-bot[rate-limiter,webhooks]==20.2
openai==0.27.4
tiktoken>=0.3.0
PyYAML==6.0
pymongo==4.3.3
python-dotenv==1.0.0
pydub==0.25.1